update table structure

+++ services/db-dwh/ddl/dim_mitra_bisnis.sql
@@ -1,6 +1,6 @@
 -- Table
 CREATE TABLE dim_mitra_bisnis (
-    id INT8,
+    id SERIAL,
     id_unit_ternak INT8,
     nama_mitra_bisnis VARCHAR(100),
     kategori_mitra_bisnis VARCHAR(100),
